+++ src/components/ConfigEditor/AzureCredentials.ts
@@ -1,4 +1,4 @@
-import { DataSourceSettings } from '@grafana/data';
+import { DataSourceSettings, SelectableValue } from '@grafana/data';
 import { AdxDataSourceOptions, AdxDataSourceSecureOptions } from 'types';
 import { getCredentials } from './AzureCredentialsConfig';
 
@@ -7,6 +7,12 @@ export enum AzureCloud {
   China = 'AzureChinaCloud',
   USGovernment = 'AzureUSGovernment',
 }
+
+export const KnownAzureClouds = [
+  { value: AzureCloud.Public, label: 'Azure' },
+  { value: AzureCloud.China, label: 'Azure China' },
+  { value: AzureCloud.USGovernment, label: 'Azure US Government' },
+] as SelectableValue[];
 
 export function isCredentialsComplete(
   options: DataSourceSettings<AdxDataSourceOptions, AdxDataSourceSecureOptions>
